+++ src/median_filter.cu
@@ -1,0 +1,25 @@
+/**
+    This file is part of sgm. (https://github.com/dhernandez0/sgm).
+
+    Copyright (c) 2016 Daniel Hernandez Juarez.
+
+    sgm is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    sgm is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with sgm.  If not, see <http://www.gnu.org/licenses/>.
+
+**/
+
+#include "median_filter.h"
+
+__global__ void MedianFilter3x3(const uint8_t* __restrict__ d_input, uint8_t* __restrict__ d_out, const uint32_t rows, const uint32_t cols) {
+	MedianFilter<3>(d_input, d_out, rows, cols);
+}
